[BE-3] Bugfix - fix endpoint for remove account

+++ controllers/SettingsController.js
@@ -13,10 +13,15 @@ import PersonalTools from "../models/PersonalTools.js";
 export const removeAccount = async (req, res) => {
   try {
     const userId = req.params.userId;
+    const userName = req.query.userName;
     const user = await UserModel.findById(userId);
 
     if (!user) {
       return getError(res, 404, { message: 'User not found!' });
+    }
+
+    if (user.userName !== userName) {
+      return getError(res, 403, { message: 'Username is not correct!' });
     }
 
     await UserModel.findByIdAndDelete(userId);
